refactor: extract types into declaration files
