+++ lib/components/key/Key.d.ts
@@ -1,0 +1,5 @@
+export type KeyProps = {
+  children: React.ReactNode;
+  className?: string;
+  size?: 'small' | 'middle' | 'large';
+}
