+++ lib/components/search-modal/SearchModal.d.ts
@@ -1,0 +1,7 @@
+import type { Config } from '../../Search.d';
+
+export type SearchModalProps = {
+  config: Config;
+  isOpen: boolean;
+  onClose?: any;
+}
