+++ lib/components/search-modal/SearchModal.tsx
@@ -1,7 +1,6 @@
 import { ChangeEvent, FormEvent, useEffect, useRef, useState } from 'react';
 import Modal from 'react-modal';
-import { getLocalStorageItem, setLocalStorageItem } from '../../helpers/localstorage';
-import type { Config } from '../../Search';
+import { getLocalStorageItem, setLocalStorageItem } from '../../helpers/local-storage';
 import BackArrowIcon from '../icons/BackArrowIcon';
 import MagicIcon from '../icons/MagicIcon';
 import SearchIcon from '../icons/SearchIcon';
@@ -9,33 +8,16 @@ import Key from '../key/Key';
 import Answer from './components/answer/Answer';
 import DocsList from './components/docs-list/DocsList';
 import Footer from './components/footer/Footer';
-import { AnswerType, getAnswers } from './services/answers';
-import { DocsType, DocType, getDocs } from './services/docs';
+import { getAnswers } from './services/answers';
+import { getDocs } from './services/docs';
+import { formatHits } from './helpers/format';
 import classes from './SearchModal.module.css';
-
-function formatHits(hits: any) {
-  return hits.map((hit: any) => {
-    const { objectID, hierarchy, content, type, url } = hit.document;
-    const hierarchyValues = Object.values(hierarchy).filter(Boolean);
-
-    return {
-      _id: objectID,
-      title: hierarchyValues[hierarchyValues.length - 1],
-      descrption: content,
-      type,
-      url
-    };
-  });
-}
+import type { SearchModalProps } from './SearchModal.d';
+import type { AnswerType } from './services/answers.d';
+import type { DocsType, DocType } from './services/docs.d';
 
 const INITIAL_ANSWER = { _id: '', search: '', answer: '', sources: [] };
 const INITIAL_DOCS: DocsType = [];
-
-type SearchModalProps = {
-  config: Config;
-  isOpen: boolean;
-  onClose?: any;
-}
 
 function SearchModal({ config, isOpen, onClose }: SearchModalProps) {
   const inputRef = useRef(null);
@@ -52,9 +34,9 @@ function SearchModal({ config, isOpen, onClose }: SearchModalProps) {
   }
 
   async function handleSearchDocs(event: ChangeEvent<HTMLInputElement>) {
-    if (config.typesense) {
+    if (config.docSearch) {
       try {
-        const { hits } = await getDocs({ config: config.typesense, search: event.target.value });
+        const { hits } = await getDocs({ config: config.docSearch, search: event.target.value });
         setDocs(formatHits(hits));
       } catch (error) {
         console.error('Search docs', error);
@@ -78,7 +60,7 @@ function SearchModal({ config, isOpen, onClose }: SearchModalProps) {
       const search = formValues.search as string;
 
       if (search) {
-        const data = await getAnswers({ accessToken: config.enhancedocs.accessToken, search });
+        const data = await getAnswers({ accessToken: config.enhancedSearch.accessToken, search });
         setAnswer({
           search,
           _id: data._id,
@@ -144,7 +126,7 @@ function SearchModal({ config, isOpen, onClose }: SearchModalProps) {
 
       <div className={classes.EnhancedSearch__SearchModal__InnerBody}>
         <Answer
-          accessToken={config.enhancedocs.accessToken}
+          accessToken={config.enhancedSearch.accessToken}
           answer={answer}
           loading={loadingAnswer}
         />
