+++ lib/components/search-modal/components/answer/Answer.d.ts
@@ -1,0 +1,7 @@
+import type { AnswerType } from '../../services/answers.d';
+
+export type AnswerProps = {
+  accessToken: string;
+  answer: AnswerType;
+  loading?: boolean;
+}
