+++ lib/components/search-modal/components/answer/components/feedback/Feedback.module.css
@@ -35,6 +35,44 @@
   box-shadow: none;
 }
 
+.EnhancedSearch__DotStretching {
+  position: relative;
+  width: 8px;
+  height: 8px;
+  border-radius: 5px;
+  background-color: var(--color-primary-base);
+  color: var(--color-primary-base);
+  transform: scale(1.25, 1.25);
+  margin: 0px 16px;
+  animation: stretching 1s infinite ease-in;
+}
+
+.EnhancedSearch__DotStretching::before,
+.EnhancedSearch__DotStretching::after {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+}
+
+.EnhancedSearch__DotStretching::before {
+  width: 8px;
+  height: 8px;
+  border-radius: 5px;
+  background-color: var(--color-primary-base);
+  color: var(--color-primary-base);
+  animation: stretching-before 1s infinite ease-in;
+}
+
+.EnhancedSearch__DotStretching::after {
+  width: 8px;
+  height: 8px;
+  border-radius: 5px;
+  background-color: var(--color-primary-base);
+  color: var(--color-primary-base);
+  animation: stretching-after 1s infinite ease-in;
+}
+
 @keyframes fadeInDown {
   from {
     opacity: 0;
@@ -44,5 +82,41 @@
   to {
     opacity: 1;
     transform: translate3d(0, 0, 0);
+  }
+}
+
+@keyframes stretching {
+  0% {
+    transform: scale(1.25, 1.25);
+  }
+  50%, 60% {
+    transform: scale(0.8, 0.8);
+  }
+  100% {
+    transform: scale(1.25, 1.25);
+  }
+}
+
+@keyframes stretching-before {
+  0% {
+    transform: translate(0) scale(0.7, 0.7);
+  }
+  50%, 60% {
+    transform: translate(-12px) scale(1, 1);
+  }
+  100% {
+    transform: translate(0) scale(0.7, 0.7);
+  }
+}
+
+@keyframes stretching-after {
+  0% {
+    transform: translate(0) scale(0.7, 0.7);
+  }
+  50%, 60% {
+    transform: translate(12px) scale(1, 1);
+  }
+  100% {
+    transform: translate(0) scale(0.7, 0.7);
   }
 }
