--- lib/components/search-modal/components/answer/components/feedback/components/dot-stretching/DotStretching.tsx
@@ -1,7 +1,0 @@
-import classes from './DotStretching.module.css';
-
-function DotStretching() {
-  return <div className={classes.EnhancedSearch__DotStretching} />;
-}
-
-export default DotStretching;
