+++ lib/components/search-modal/components/footer/Footer.tsx
@@ -1,5 +1,4 @@
 import Key from '../../../key/Key';
-import Logo from './components/logo/Logo';
 import classes from './Footer.module.css';
 
 function Footer() {
@@ -21,7 +20,12 @@ function Footer() {
           target="_blank"
           rel="noreferrer noopener"
         >
-          <Logo className={classes.EnhancedSearch__SearchModal__FooterLogo} />
+          <svg className={classes.EnhancedSearch__SearchModal__FooterLogo} viewBox="0 0 2022 305" fill="none" xmlns="http://www.w3.org/2000/svg">
+            <path d="M11 80L43.5 203.5M146.5 44V80M64 47V57M76.5 34H86.5M64 12V22M51 34H39.5M212.5 143V133M223.5 155.5H233.5M212.5 169.5V178M199.5 155.5H191M266.5 200L297 82M146.5 18.5V24.5M201.5 51.5L208 70.5L220.5 54H241.5L230 37.5L236.5 18.5L216.5 24.5L199.5 12V31.5L181.5 44L201.5 51.5ZM102 73.5L90.5 97L66.5 101.5L83 119L79.5 146L102 133L126.5 146L122.5 119L141 101.5L115.5 99L102 73.5Z" stroke="black" strokeWidth="12" strokeLinecap="round" />
+            <path d="M138.5 283.5H12.5L25.5 267.135M138.5 283.5C139.5 288.5 144 298.5 154 298.5C164 298.5 168.167 288.5 169 283.5M138.5 283.5L154 269M169 283.5H207M169 283.5L154 269M154 269V264.5M154 225.5C151.833 221.333 144 212.9 130 212.5C116 212.1 74.1667 224 55 230L25.5 267.135M154 225.5C156.167 221.167 164.2 212.5 179 212.5C193.8 212.5 234.833 224.167 253.5 230L282.5 266.081M154 225.5V260V264.5M253.5 283.5H296.5L282.5 266.081M154 264.5C158.833 257 172.8 242.3 190 243.5C207.2 244.7 258.833 259.054 282.5 266.081M154 264.5C148.667 256.333 133.1 240.7 113.5 243.5C93.9 246.3 46.6667 260.424 25.5 267.135M228.5 283.5H231" stroke="#3763EB" strokeWidth="12" strokeLinecap="round" />
+            <path d="M1427.57 63.0449H1493.61C1508.5 63.0449 1519.85 64.1842 1527.66 66.4629C1538.16 69.5553 1547.15 75.0485 1554.64 82.9424C1562.13 90.8363 1567.83 100.521 1571.73 111.995C1575.64 123.388 1577.59 137.467 1577.59 154.231C1577.59 168.961 1575.76 181.657 1572.1 192.317C1567.62 205.338 1561.23 215.877 1552.93 223.934C1546.67 230.037 1538.2 234.798 1527.54 238.216C1519.57 240.739 1508.91 242 1495.56 242H1427.57V63.0449ZM1463.7 93.3184V211.849H1490.68C1500.77 211.849 1508.05 211.279 1512.53 210.14C1518.39 208.675 1523.23 206.193 1527.05 202.693C1530.96 199.194 1534.13 193.457 1536.58 185.481C1539.02 177.425 1540.24 166.479 1540.24 152.645C1540.24 138.81 1539.02 128.19 1536.58 120.784C1534.13 113.379 1530.72 107.601 1526.32 103.45C1521.93 99.2998 1516.35 96.4922 1509.6 95.0273C1504.55 93.888 1494.66 93.3184 1479.93 93.3184H1463.7ZM1600.17 175.35C1600.17 163.956 1602.98 152.929 1608.6 142.269C1614.21 131.608 1622.15 123.47 1632.4 117.854C1642.74 112.239 1654.25 109.432 1666.95 109.432C1686.56 109.432 1702.63 115.82 1715.16 128.597C1727.7 141.292 1733.96 157.365 1733.96 176.814C1733.96 196.427 1727.62 212.703 1714.92 225.643C1702.31 238.501 1686.4 244.93 1667.19 244.93C1655.31 244.93 1643.96 242.244 1633.13 236.873C1622.39 231.502 1614.21 223.649 1608.6 213.313C1602.98 202.897 1600.17 190.242 1600.17 175.35ZM1635.33 177.181C1635.33 190.039 1638.38 199.886 1644.49 206.722C1650.59 213.558 1658.12 216.976 1667.07 216.976C1676.02 216.976 1683.51 213.558 1689.53 206.722C1695.63 199.886 1698.68 189.957 1698.68 176.937C1698.68 164.241 1695.63 154.476 1689.53 147.64C1683.51 140.804 1676.02 137.386 1667.07 137.386C1658.12 137.386 1650.59 140.804 1644.49 147.64C1638.38 154.476 1635.33 164.323 1635.33 177.181ZM1873.98 150.691L1840.16 156.795C1839.02 150.04 1836.42 144.954 1832.35 141.536C1828.36 138.118 1823.16 136.409 1816.73 136.409C1808.18 136.409 1801.35 139.38 1796.22 145.32C1791.17 151.18 1788.65 161.027 1788.65 174.861C1788.65 190.242 1791.21 201.106 1796.34 207.454C1801.55 213.802 1808.51 216.976 1817.21 216.976C1823.73 216.976 1829.06 215.145 1833.21 211.482C1837.36 207.739 1840.29 201.351 1842 192.317L1875.69 198.055C1872.19 213.517 1865.47 225.195 1855.54 233.089C1845.62 240.983 1832.31 244.93 1815.63 244.93C1796.67 244.93 1781.53 238.948 1770.22 226.985C1758.99 215.022 1753.37 198.462 1753.37 177.303C1753.37 155.9 1759.03 139.257 1770.34 127.376C1781.65 115.413 1796.95 109.432 1816.24 109.432C1832.03 109.432 1844.56 112.85 1853.84 119.686C1863.19 126.44 1869.91 136.775 1873.98 150.691ZM1888.02 205.013L1922.44 199.764C1923.9 206.437 1926.87 211.523 1931.35 215.022C1935.83 218.44 1942.09 220.149 1950.15 220.149C1959.02 220.149 1965.69 218.522 1970.17 215.267C1973.18 212.988 1974.69 209.936 1974.69 206.111C1974.69 203.507 1973.87 201.351 1972.24 199.642C1970.54 198.014 1966.71 196.508 1960.77 195.125C1933.1 189.021 1915.56 183.447 1908.16 178.401C1897.9 171.403 1892.78 161.678 1892.78 149.227C1892.78 137.996 1897.21 128.556 1906.08 120.906C1914.95 113.257 1928.71 109.432 1947.34 109.432C1965.08 109.432 1978.27 112.321 1986.89 118.099C1995.52 123.877 2001.46 132.422 2004.71 143.733L1972.37 149.715C1970.98 144.669 1968.34 140.804 1964.43 138.118C1960.61 135.433 1955.11 134.09 1947.95 134.09C1938.92 134.09 1932.45 135.351 1928.54 137.874C1925.94 139.664 1924.64 141.984 1924.64 144.832C1924.64 147.273 1925.78 149.349 1928.05 151.058C1931.15 153.336 1941.81 156.551 1960.04 160.701C1978.35 164.852 1991.12 169.938 1998.37 175.96C2005.53 182.063 2009.11 190.568 2009.11 201.473C2009.11 213.354 2004.15 223.567 1994.22 232.112C1984.29 240.657 1969.6 244.93 1950.15 244.93C1932.49 244.93 1918.49 241.349 1908.16 234.188C1897.9 227.026 1891.19 217.301 1888.02 205.013Z" fill="#3763EB" />
+            <path d="M380.188 242V63.0449H512.879V93.3184H416.321V132.991H506.165V163.143H416.321V211.849H516.297V242H380.188ZM664.612 242H630.311V175.838C630.311 161.84 629.578 152.807 628.113 148.738C626.648 144.588 624.248 141.373 620.911 139.095C617.656 136.816 613.709 135.677 609.07 135.677C603.13 135.677 597.799 137.304 593.079 140.56C588.359 143.815 585.104 148.128 583.313 153.499C581.604 158.87 580.75 168.799 580.75 183.284V242H546.448V112.361H578.309V131.404C589.62 116.756 603.862 109.432 621.033 109.432C628.602 109.432 635.519 110.815 641.785 113.582C648.051 116.268 652.771 119.726 655.945 123.958C659.201 128.19 661.438 132.991 662.659 138.362C663.961 143.733 664.612 151.424 664.612 161.434V242ZM733.704 63.0449V128.841C744.772 115.901 757.996 109.432 773.377 109.432C781.271 109.432 788.392 110.896 794.739 113.826C801.087 116.756 805.848 120.499 809.021 125.057C812.277 129.614 814.474 134.66 815.613 140.193C816.834 145.727 817.444 154.313 817.444 165.95V242H783.143V173.519C783.143 159.928 782.492 151.302 781.189 147.64C779.887 143.978 777.568 141.089 774.231 138.973C770.976 136.775 766.867 135.677 761.902 135.677C756.206 135.677 751.119 137.06 746.644 139.827C742.168 142.594 738.872 146.785 736.756 152.4C734.721 157.934 733.704 166.154 733.704 177.059V242H699.402V63.0449H733.704ZM877.991 151.912L846.863 146.297C850.363 133.764 856.385 124.487 864.93 118.465C873.475 112.443 886.17 109.432 903.016 109.432C918.315 109.432 929.708 111.263 937.195 114.925C944.682 118.506 949.931 123.104 952.942 128.719C956.035 134.253 957.581 144.466 957.581 159.358L957.215 199.397C957.215 210.791 957.744 219.214 958.802 224.666C959.941 230.037 962.016 235.815 965.027 242H931.092C930.197 239.721 929.098 236.344 927.796 231.868C927.226 229.834 926.819 228.491 926.575 227.84C920.716 233.536 914.45 237.809 907.776 240.657C901.103 243.506 893.982 244.93 886.414 244.93C873.068 244.93 862.529 241.308 854.798 234.065C847.148 226.823 843.323 217.667 843.323 206.6C843.323 199.275 845.073 192.765 848.572 187.068C852.072 181.29 856.954 176.896 863.221 173.885C869.568 170.792 878.683 168.107 890.564 165.828C906.596 162.817 917.705 160.009 923.89 157.405V153.987C923.89 147.396 922.262 142.716 919.007 139.949C915.752 137.101 909.607 135.677 900.574 135.677C894.471 135.677 889.71 136.897 886.292 139.339C882.874 141.699 880.107 145.89 877.991 151.912ZM923.89 179.744C919.495 181.209 912.537 182.959 903.016 184.993C893.494 187.028 887.269 189.021 884.339 190.975C879.863 194.148 877.625 198.177 877.625 203.06C877.625 207.861 879.415 212.011 882.996 215.511C886.577 219.01 891.134 220.76 896.668 220.76C902.853 220.76 908.753 218.725 914.368 214.656C918.519 211.564 921.245 207.78 922.547 203.304C923.442 200.374 923.89 194.799 923.89 186.58V179.744ZM1109.44 242H1075.13V175.838C1075.13 161.84 1074.4 152.807 1072.94 148.738C1071.47 144.588 1069.07 141.373 1065.74 139.095C1062.48 136.816 1058.53 135.677 1053.89 135.677C1047.95 135.677 1042.62 137.304 1037.9 140.56C1033.18 143.815 1029.93 148.128 1028.14 153.499C1026.43 158.87 1025.57 168.799 1025.57 183.284V242H991.272V112.361H1023.13V131.404C1034.44 116.756 1048.69 109.432 1065.86 109.432C1073.43 109.432 1080.34 110.815 1086.61 113.582C1092.88 116.268 1097.6 119.726 1100.77 123.958C1104.02 128.19 1106.26 132.991 1107.48 138.362C1108.79 143.733 1109.44 151.424 1109.44 161.434V242ZM1257.39 150.691L1223.57 156.795C1222.43 150.04 1219.83 144.954 1215.76 141.536C1211.77 138.118 1206.56 136.409 1200.13 136.409C1191.59 136.409 1184.75 139.38 1179.63 145.32C1174.58 151.18 1172.06 161.027 1172.06 174.861C1172.06 190.242 1174.62 201.106 1179.75 207.454C1184.96 213.802 1191.92 216.976 1200.62 216.976C1207.13 216.976 1212.46 215.145 1216.61 211.482C1220.76 207.739 1223.69 201.351 1225.4 192.317L1259.09 198.055C1255.6 213.517 1248.88 225.195 1238.95 233.089C1229.02 240.983 1215.72 244.93 1199.04 244.93C1180.07 244.93 1164.94 238.948 1153.63 226.985C1142.4 215.022 1136.78 198.462 1136.78 177.303C1136.78 155.9 1142.44 139.257 1153.75 127.376C1165.06 115.413 1180.36 109.432 1199.65 109.432C1215.43 109.432 1227.97 112.85 1237.24 119.686C1246.6 126.44 1253.32 136.775 1257.39 150.691ZM1358.58 200.74L1392.76 206.478C1388.37 219.01 1381.41 228.572 1371.89 235.164C1362.45 241.674 1350.61 244.93 1336.37 244.93C1313.82 244.93 1297.14 237.565 1286.32 222.835C1277.77 211.035 1273.5 196.142 1273.5 178.157C1273.5 156.673 1279.11 139.868 1290.34 127.742C1301.58 115.535 1315.78 109.432 1332.95 109.432C1352.23 109.432 1367.45 115.82 1378.6 128.597C1389.75 141.292 1395.08 160.783 1394.59 187.068H1308.66C1308.9 197.241 1311.67 205.175 1316.96 210.872C1322.25 216.487 1328.84 219.295 1336.73 219.295C1342.1 219.295 1346.62 217.83 1350.28 214.9C1353.94 211.971 1356.71 207.251 1358.58 200.74ZM1360.54 166.072C1360.29 156.144 1357.73 148.616 1352.84 143.489C1347.96 138.281 1342.02 135.677 1335.02 135.677C1327.54 135.677 1321.35 138.403 1316.47 143.855C1311.58 149.308 1309.18 156.714 1309.27 166.072H1360.54Z" fill="black" />
+          </svg>
         </a>
       </div>
     </footer>
