+++ lib/components/search-modal/services/answers.d.ts
@@ -1,0 +1,17 @@
+export type AnswerType = {
+  _id: string;
+  search: string;
+  answer: string;
+  sources: Array<string>;
+};
+
+export type GetAnswers = {
+  accessToken: string;
+  search: string;
+}
+
+export type AnswerFeedbackType = {
+  accessToken: string;
+  answerId: string;
+  usefulFeedback: boolean;
+}
