+++ lib/components/search-modal/services/answers.ts
@@ -1,25 +1,8 @@
 import { Get, Post } from './instance';
-
-export type AnswerType = {
-  _id: string;
-  search: string;
-  answer: string;
-  sources: Array<string>;
-};
-
-type GetAnswers = {
-  accessToken: string;
-  search: string;
-}
+import type { AnswerType, GetAnswers, AnswerFeedbackType } from './answers.d';
 
 export function getAnswers({ accessToken, search }: GetAnswers): Promise<AnswerType> {
   return Get(`/questions/ask?q=${search}`, { headers: { Authorization: `Bearer ${accessToken}` } });
-}
-
-type AnswerFeedbackType = {
-  accessToken: string;
-  answerId: string;
-  usefulFeedback: boolean;
 }
 
 export function answerFeedback({ accessToken, answerId, usefulFeedback } : AnswerFeedbackType) {
