--- lib/helpers/classnames.ts
@@ -1,5 +1,0 @@
-function classNames(...classes: Array<string | boolean | undefined>) {
-  return classes.filter(Boolean).join(' ');
-}
-
-export default classNames;
