+++ lib/helpers/styles.ts
@@ -1,0 +1,3 @@
+export function classNames(...classes: Array<string | boolean | undefined>) {
+  return classes.filter(Boolean).join(' ');
+}
